Improve docs
