+++ src/mlpack/bindings/java/print_doc_functions_impl.hpp
@@ -1,3 +1,9 @@
+/**
+ * @file print_doc_functions_impl.hpp
+ * @author Vasyl Teliman
+ *
+ * Implementation of functions to generate documentation
+ */
 #ifndef MLPACK_BINDINGS_JAVA_PRINT_DOC_FUNCTIONS_IMPL_HPP
 #define MLPACK_BINDINGS_JAVA_PRINT_DOC_FUNCTIONS_IMPL_HPP
 
@@ -7,9 +13,15 @@ namespace mlpack {
 namespace bindings {
 namespace java {
 
+/**
+ * Base case for template recursion below
+ */
 inline void PrintInputCallParam(std::ostream&)
 {}
 
+/**
+ * Generate input arguments for function call
+ */
 template <typename T, typename... Args>
 void PrintInputCallParam(std::ostream& os, const std::string& name,
     const T& value, const Args&... args)
@@ -46,9 +58,15 @@ void PrintInputCallParam(std::ostream& os, const std::string& name,
   PrintInputCallParam(os, args...);
 }
 
+/**
+ * Base case for template recursion below
+ */
 inline void PrintOutputCallParam(std::ostream&)
 {}
 
+/**
+ * Print output values for a function call
+ */
 template <typename T, typename... Args>
 void PrintOutputCallParam(std::ostream& os, const std::string& name,
     const T& /* value */, const Args&... args)
@@ -68,11 +86,17 @@ void PrintOutputCallParam(std::ostream& os, const std::string& name,
   PrintOutputCallParam(os, args...);
 }
 
+/**
+ * Generate documentation for string parameter
+ */
 inline std::string ParamString(const std::string& paramName)
 {
   return "{@code " + paramName + '}';
 }
 
+/**
+ * Generate documentation for value parameter
+ */
 template<typename T>
 inline std::string PrintValue(const T& value, bool quotes)
 {
@@ -83,16 +107,25 @@ inline std::string PrintValue(const T& value, bool quotes)
   return oss.str();
 }
 
+/**
+ * Generate documentation for dataset parameter
+ */
 inline std::string PrintDataset(const std::string& datasetName)
 {
   return "{@code " + datasetName + '}';
 }
 
+/**
+ * Generate documentation for model parameter
+ */
 inline std::string PrintModel(const std::string& modelName)
 {
   return "{@code " + modelName + '}';
 }
 
+/**
+ * Generate documentation string for function calls
+ */
 template<typename... Args>
 std::string ProgramCall(const std::string& /* programName */, Args... args)
 {
@@ -119,6 +152,9 @@ std::string ProgramCall(const std::string& /* programName */, Args... args)
   return oss.str();
 }
 
+/**
+ * Generate function call without arguments
+ */
 inline std::string ProgramCall(const std::string& /* programName */)
 {
   std::ostringstream oss;
@@ -133,11 +169,17 @@ inline std::string ProgramCall(const std::string& /* programName */)
   return oss.str();
 }
 
+/**
+ * Ignore check only for output parameters
+ */
 inline bool IgnoreCheck(const std::string& paramName)
 {
   return !CLI::Parameters()[paramName].input;
 }
 
+/**
+ * Ignore check if at least one parameter is output parameter
+ */
 inline bool IgnoreCheck(const std::vector<std::string>& constraints)
 {
   for (size_t i = 0; i < constraints.size(); ++i)
@@ -149,6 +191,9 @@ inline bool IgnoreCheck(const std::vector<std::string>& constraints)
   return false;
 }
 
+/**
+ * Ignore check if at least one parameter is output parameter
+ */
 inline bool IgnoreCheck(
   const std::vector<std::pair<std::string, bool>>& constraints,
   const std::string& paramName)
